Added Workspace class to re-use allocations in repeated trend fits.
This is useful for re-using memory when fitting across multiple blocks.
Removed the need for dedicated buffers for the fitted values and
residuals, we just shift things around on the same buffer now.

Renamed the fixed_width mode to minimum_width, which is a bit more
accurate given that the window expands if there's not enough points; and
pass along the number of threads to the LOWESS fitter.

Added some tests for correct behavior with mean-based filtering.